recurrect methods

+++ src/main/java/com/CyberMallBackEnd/CyberMallBackEnd/dto/CartDto.java
@@ -7,13 +7,13 @@ import java.util.List;
 
 @Data
 public class CartDto {
-    private int customerId; // Keep customer ID at the root
-    private int cartId;
+    private Integer customerId; // Keep customer ID at the root
+    private Integer cartId;
     private List<ProductDetail> productDetails; // Fixed variable name to follow Java conventions
 
     @Data
     public static class ProductDetail {
-        private Long productId;
+        private Integer productId;
         private String title;
         private BigDecimal price; // Use BigDecimal for monetary values
         private int quantity;
