+++ src/main/java/com/CyberMallBackEnd/CyberMallBackEnd/repository/ProductRepository.java
@@ -3,5 +3,5 @@ package com.CyberMallBackEnd.CyberMallBackEnd.repository;
 import com.CyberMallBackEnd.CyberMallBackEnd.entity.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductRepository extends JpaRepository<Product, Long> {
+public interface ProductRepository extends JpaRepository<Product, Integer> {
 }
